@@ -5,16 +5,17 @@ publish_to: none
 version: 2.3.2+36
 
 environment:
-  sdk: ">=3.5.0 <4.0.0"
+  sdk: ">=3.6.0 <4.0.0"
 
 dependencies:
   collection: any
   csv: ^6.0.0
+  dart_mappable: ^4.3.0
   dart_rss:
     git:
       url: https://github.com/Chonli/webfeed.git
       ref: feature/geo-rss-2
-  fl_chart: ^0.69.0
+  fl_chart: ^0.70.1
   flutter:
     sdk: flutter
   flutter_hooks: ^0.20.5
@@ -22,40 +23,39 @@ dependencies:
   flutter_launcher_icons: ^0.14.1
   flutter_map: ^7.0.2
   flutter_map_marker_cluster: ^1.4.0
-  flutter_sticky_header: ^0.6.5
-  freezed_annotation: ^2.4.1
+  flutter_sticky_header: ^0.7.0
   geolocator: ^13.0.1
   go_router: ^14.2.0
   hooks_riverpod: ^2.5.1
   http: ^1.2.1
   intl: any
-  json_annotation: ^4.9.0
   latlong2: ^0.9.0
   logging: ^1.2.0
-  package_info: ^2.0.2
+  package_info_plus: ^8.1.1
   path: ^1.9.0
+  path_provider: ^2.1.5
   pdfx: ^2.8.0
   riverpod_annotation: ^2.3.5
-  share: ^2.0.4
+  sembast: ^3.8.0+1
+  share_plus: ^10.1.2
   shared_preferences: ^2.2.2
+  sliver_tools: ^0.2.12
   smooth_page_indicator: ^1.2.0
-  sqflite: ^2.3.2
   url_launcher: ^6.3.0
 
 dev_dependencies:
   build_runner: ^2.4.13
-  build_runner_core: ^7.3.2
-  custom_lint: ^0.6.7
+  custom_lint: ^0.7.0
+  dart_mappable_builder: ^4.3.0
   flutter_lints: ^5.0.0
   flutter_test:
     sdk: flutter
-  freezed: ^2.5.7
+  freezed: ^2.4.7
   json_serializable: ^6.8.0
-  riverpod_generator: ^2.4.3
-  very_good_analysis: ^6.0.0
-
-dependency_overrides:
-  win32: ^5.5.4
+  mocktail: ^1.0.4
+  riverpod_generator: ^2.3.11
+  riverpod_lint: ^2.3.13
+  very_good_analysis: ^7.0.0
 
 flutter:
   uses-material-design: true
