@@ -6,7 +6,7 @@ part of 'location.dart';
 // RiverpodGenerator
 // **************************************************************************
 
-String _$userLocationHash() => r'a572545ff042d094f80a6e0a1061ae633f6e1fe1';
+String _$userLocationHash() => r'e7cc84cfbc66f5cb9781034f1bedb5583105f196';
 
 /// See also [UserLocation].
 @ProviderFor(UserLocation)
@@ -22,4 +22,4 @@ final userLocationProvider =
 
 typedef _$UserLocation = AutoDisposeAsyncNotifier<LatLng?>;
 // ignore_for_file: type=lint
-// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package
