@@ -1,5 +1,5 @@
-import 'package:latlong2/latlong.dart';
 import 'package:snow_weather_info/model/avalanche_bulletin.dart';
+import 'package:snow_weather_info/model/coordinate.dart';
 import 'package:snow_weather_info/model/mountain.dart';
 import 'package:snow_weather_info/model/station.dart';
 
@@ -8,183 +8,183 @@ abstract class ConstantDatalist {
 
   static final listNivose = <Nivose>[
     //alpes nord
-    Nivose(
+    const Nivose(
       'Aiguilles rouges',
-      const LatLng(45.986125, 6.895937),
+      Coordinate(latitude: 45.986125, longitude: 6.895937),
       2330,
       'AIGRG',
     ),
-    Nivose(
+    const Nivose(
       'Bellecote',
-      const LatLng(45.48881, 6.77163),
+      Coordinate(latitude: 45.48881, longitude: 6.77163),
       3000,
       'BELLE',
     ),
-    Nivose(
+    const Nivose(
       'Le Chevril',
-      const LatLng(45.50179, 6.95604),
+      Coordinate(latitude: 45.50179, longitude: 6.95604),
       2560,
       'CHEVR',
     ),
-    Nivose(
+    const Nivose(
       'Bonneval',
-      const LatLng(45.35175, 7.05132),
+      Coordinate(latitude: 45.35175, longitude: 7.05132),
       2720,
       'BONNE',
     ),
-    Nivose(
+    const Nivose(
       'Les rochilles',
-      const LatLng(45.08467, 6.47127),
+      Coordinate(latitude: 45.08467, longitude: 6.47127),
       2450,
       'ROCHI',
     ),
-    Nivose(
+    const Nivose(
       'Allant',
-      const LatLng(45.66345, 6.21195),
+      Coordinate(latitude: 45.66345, longitude: 6.21195),
       1630,
       'ALLAN',
     ),
-    Nivose(
+    const Nivose(
       'Grande Parei',
-      const LatLng(45.61463, 6.64750),
+      Coordinate(latitude: 45.61463, longitude: 6.64750),
       2240,
       'PAREI',
     ),
-    Nivose(
+    const Nivose(
       'Col de Porte',
-      const LatLng(45.2952742, 5.7651929),
+      Coordinate(latitude: 45.2952742, longitude: 5.7651929),
       1325,
       'COLPO',
     ),
-    Nivose(
+    const Nivose(
       'St Hilaire',
-      const LatLng(45.31449, 5.86462),
+      Coordinate(latitude: 45.31449, longitude: 5.86462),
       1700,
       'STHIL',
     ),
-    Nivose(
+    const Nivose(
       'Aigleton',
-      const LatLng(45.2311598, 6.0381140),
+      Coordinate(latitude: 45.2311598, longitude: 6.0381140),
       2240,
       'AIGLE',
     ),
-    Nivose(
+    const Nivose(
       'Le Gua',
-      const LatLng(45.01517, 5.58951),
+      Coordinate(latitude: 45.01517, longitude: 5.58951),
       1600,
       'LEGUA',
     ),
-    Nivose(
+    const Nivose(
       'Les Ecrins',
-      const LatLng(44.93681, 6.34602),
+      Coordinate(latitude: 44.93681, longitude: 6.34602),
       2940,
       'ECRIN',
     ),
-    Nivose(
+    const Nivose(
       'La Meije',
-      const LatLng(45.01255, 6.26485),
+      Coordinate(latitude: 45.01255, longitude: 6.26485),
       3100,
       'MEIJE',
     ),
-    Nivose(
+    const Nivose(
       'Galibier',
-      const LatLng(45.05733, 6.37739),
+      Coordinate(latitude: 45.05733, longitude: 6.37739),
       2559,
       'GALIB',
     )
     //alpes sud
     ,
-    Nivose(
+    const Nivose(
       'Orcières',
-      const LatLng(44.71857, 6.33378),
+      Coordinate(latitude: 44.71857, longitude: 6.33378),
       2294,
       'ORCIE',
     ),
-    Nivose(
+    const Nivose(
       'Col Agnel',
-      const LatLng(44.68873, 6.97625),
+      Coordinate(latitude: 44.68873, longitude: 6.97625),
       2630,
       'AGNEL',
     ),
-    Nivose(
+    const Nivose(
       'Restefond',
-      const LatLng(44.34205, 6.79871),
+      Coordinate(latitude: 44.34205, longitude: 6.79871),
       2550,
       'RESTE',
     ),
-    Nivose(
+    const Nivose(
       'Millefonts',
-      const LatLng(44.11819, 7.19226),
+      Coordinate(latitude: 44.11819, longitude: 7.19226),
       2430,
       'MILLE',
     ),
-    Nivose(
+    const Nivose(
       'Parpaillon',
-      const LatLng(44.49522, 6.63650),
+      Coordinate(latitude: 44.49522, longitude: 6.63650),
       2545,
       'PARPA',
     )
     //corse
     ,
-    Nivose(
+    const Nivose(
       'Sponde',
-      const LatLng(42.32448, 8.88478),
+      Coordinate(latitude: 42.32448, longitude: 8.88478),
       1980,
       'SPOND',
     ),
-    Nivose(
+    const Nivose(
       'Maniccia',
-      const LatLng(42.20837, 9.04888),
+      Coordinate(latitude: 42.20837, longitude: 9.04888),
       2360,
       'MANIC',
     )
     //pyrennée
     ,
-    Nivose(
+    const Nivose(
       'Maupas',
-      const LatLng(42.71428, 0.54995),
+      Coordinate(latitude: 42.71428, longitude: 0.54995),
       2430,
       'MAUPA',
     ),
-    Nivose(
+    const Nivose(
       'Port d`Aula',
-      const LatLng(42.76993, 1.11493),
+      Coordinate(latitude: 42.76993, longitude: 1.11493),
       2140,
       'PAULA',
     ),
-    Nivose(
+    const Nivose(
       'Canigou',
-      const LatLng(42.53322, 2.46096),
+      Coordinate(latitude: 42.53322, longitude: 2.46096),
       2150,
       'CANIG',
     ),
-    Nivose(
+    const Nivose(
       'Hospitalet',
-      const LatLng(42.57778, 1.81140),
+      Coordinate(latitude: 42.57778, longitude: 1.81140),
       2293,
       'HOSPI',
     ),
-    Nivose(
+    const Nivose(
       'Puigmal',
-      const LatLng(42.38318, 2.09143),
+      Coordinate(latitude: 42.38318, longitude: 2.09143),
       2467,
       'PUIGN',
     ),
-    Nivose(
+    const Nivose(
       'Soum Couy',
-      const LatLng(42.96157, -0.71995),
+      Coordinate(latitude: 42.96157, longitude: -0.71995),
       2150,
       'SOUMC',
     ),
-    Nivose(
+    const Nivose(
       'Lac d’ardiden',
-      const LatLng(42.84828, -0.06471),
+      Coordinate(latitude: 42.84828, longitude: -0.06471),
       2445,
       'LARDI',
     ),
-    Nivose(
+    const Nivose(
       'Aiguillettes',
-      const LatLng(42.74692, 0.18128),
+      Coordinate(latitude: 42.74692, longitude: 0.18128),
       2120,
       'AIGTE',
     ),
