@@ -4,7 +4,7 @@ import 'dart:developer' show log;
 import 'package:dart_rss/dart_rss.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
-import 'package:snow_weather_info/data/sources/api_client.dart';
+import 'package:snow_weather_info/data/sources/api/api_client.dart';
 
 part 'avalanche_api.g.dart';
 
@@ -17,7 +17,7 @@ Future<AtomFeed?> apiAvalanche(Ref ref) async {
     final response =
         await client.get(Uri.parse('http://www.data-avalanche.org/feed'));
     feed = AtomFeed.parse(utf8.decode(response.bodyBytes));
-  } catch (e) {
+  } on Exception catch (e) {
     log('feed error : $e');
     throw Exception(e);
   }
