@@ -6,7 +6,7 @@ part of 'avalanche_api.dart';
 // RiverpodGenerator
 // **************************************************************************
 
-String _$apiAvalancheHash() => r'9adf64de3813a1c23dcb3e5ecc062502dd1a2997';
+String _$apiAvalancheHash() => r'f73a00238dc1c31bbf067d674ac1e551a62debf1';
 
 /// See also [apiAvalanche].
 @ProviderFor(apiAvalanche)
@@ -19,6 +19,8 @@ final apiAvalancheProvider = FutureProvider<AtomFeed?>.internal(
   allTransitiveDependencies: null,
 );
 
+@Deprecated('Will be removed in 3.0. Use Ref instead')
+// ignore: unused_element
 typedef ApiAvalancheRef = FutureProviderRef<AtomFeed?>;
 // ignore_for_file: type=lint
-// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package
