@@ -1,12 +1,11 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 part 'preferences.g.dart';
 
-const _lastStationPrefs = 'lastStationPrefs';
-const _lastStationDataPrefs = 'lastStationDataPrefs';
 const _favoritesStationPrefs = 'favoritesStationPrefs';
 
 const _favoritesSkiResortPrefs = 'favoritesSkiResortPrefs';
@@ -16,7 +15,7 @@ const _showNoDataStationPrefs = 'showNoDataStationPrefs';
 const _showClusterLayerPrefs = 'showClusterLayerPrefs';
 
 @Riverpod(keepAlive: true)
-SharedPreferences sharedPreferences(SharedPreferencesRef ref) {
+SharedPreferences sharedPreferences(Ref ref) {
   throw UnimplementedError();
 }
 
@@ -46,48 +45,6 @@ class ThemeModeSettings extends _$ThemeModeSettings {
     if (state != mode) {
       state = mode;
       _themeMode = mode;
-    }
-  }
-}
-
-@Riverpod(keepAlive: true)
-class LastStationDataSettings extends _$LastStationDataSettings {
-  @override
-  DateTime build() {
-    return _lastStationDataDate;
-  }
-
-  SharedPreferences get _preference => ref.watch(sharedPreferencesProvider);
-
-  DateTime get _lastStationDataDate => DateTime.parse(
-        _preference.getString(_lastStationDataPrefs) ?? '19700101',
-      );
-
-  void updateDate(DateTime lastDate) {
-    if (state != lastDate) {
-      state = lastDate;
-      _preference.setString(_lastStationDataPrefs, lastDate.toString());
-    }
-  }
-}
-
-@Riverpod(keepAlive: true)
-class LastStationSettings extends _$LastStationSettings {
-  @override
-  DateTime build() {
-    return _lastStationDate;
-  }
-
-  SharedPreferences get _preference => ref.watch(sharedPreferencesProvider);
-
-  DateTime get _lastStationDate => DateTime.parse(
-        _preference.getString(_lastStationPrefs) ?? '19700101',
-      );
-
-  void updateDate(DateTime lastDate) {
-    if (state.difference(lastDate) > const Duration(days: 1)) {
-      state = lastDate;
-      _preference.setString(_lastStationPrefs, lastDate.toString());
     }
   }
 }
