@@ -6,7 +6,7 @@ part of 'preferences.dart';
 // RiverpodGenerator
 // **************************************************************************
 
-String _$sharedPreferencesHash() => r'3a9f8412df34c1653d08100c9826aa2125b80f7f';
+String _$sharedPreferencesHash() => r'1a6250efdc19e86c923ceb598a77ff74d64378e6';
 
 /// See also [sharedPreferences].
 @ProviderFor(sharedPreferences)
@@ -20,6 +20,8 @@ final sharedPreferencesProvider = Provider<SharedPreferences>.internal(
   allTransitiveDependencies: null,
 );
 
+@Deprecated('Will be removed in 3.0. Use Ref instead')
+// ignore: unused_element
 typedef SharedPreferencesRef = ProviderRef<SharedPreferences>;
 String _$themeModeSettingsHash() => r'106550861a9acf66c527c4f3f77fe9c7c850ac47';
 
@@ -37,40 +39,6 @@ final themeModeSettingsProvider =
 );
 
 typedef _$ThemeModeSettings = Notifier<ThemeMode>;
-String _$lastStationDataSettingsHash() =>
-    r'6cfdbd488e1416750f361235142136088aa8fd54';
-
-/// See also [LastStationDataSettings].
-@ProviderFor(LastStationDataSettings)
-final lastStationDataSettingsProvider =
-    NotifierProvider<LastStationDataSettings, DateTime>.internal(
-  LastStationDataSettings.new,
-  name: r'lastStationDataSettingsProvider',
-  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
-      ? null
-      : _$lastStationDataSettingsHash,
-  dependencies: null,
-  allTransitiveDependencies: null,
-);
-
-typedef _$LastStationDataSettings = Notifier<DateTime>;
-String _$lastStationSettingsHash() =>
-    r'106a627e5e1a04714b24b00973c082ed155258c1';
-
-/// See also [LastStationSettings].
-@ProviderFor(LastStationSettings)
-final lastStationSettingsProvider =
-    NotifierProvider<LastStationSettings, DateTime>.internal(
-  LastStationSettings.new,
-  name: r'lastStationSettingsProvider',
-  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
-      ? null
-      : _$lastStationSettingsHash,
-  dependencies: null,
-  allTransitiveDependencies: null,
-);
-
-typedef _$LastStationSettings = Notifier<DateTime>;
 String _$favoritesStationSettingsHash() =>
     r'1558bd39ac145e4dad63815a1427b95068c6c379';
 
@@ -157,4 +125,4 @@ final showClusterLayerSettingsProvider =
 
 typedef _$ShowClusterLayerSettings = Notifier<bool>;
 // ignore_for_file: type=lint
-// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package
