@@ -6,7 +6,7 @@ part of 'view.dart';
 // RiverpodGenerator
 // **************************************************************************
 
-String _$massifFilterHash() => r'd95afe77e393622085b9724b78b6631ad1a16647';
+String _$massifFilterHash() => r'38333b2b6862a8591522d56b823077d8093a8e36';
 
 /// See also [_MassifFilter].
 @ProviderFor(_MassifFilter)
@@ -22,4 +22,4 @@ final _massifFilterProvider =
 
 typedef _$MassifFilter = AutoDisposeNotifier<List<Mountain>>;
 // ignore_for_file: type=lint
-// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package
