@@ -6,12 +6,12 @@ part of 'map_widget.dart';
 // RiverpodGenerator
 // **************************************************************************
 
-String _$currentMapLocHash() => r'a00ac1e233ca477c956a761fa3c2c75b901ab168';
+String _$currentMapLocHash() => r'd1554a09c91bf0d51437d829b641c3dff5398579';
 
 /// See also [CurrentMapLoc].
 @ProviderFor(CurrentMapLoc)
 final currentMapLocProvider =
-    AutoDisposeNotifierProvider<CurrentMapLoc, LatLng>.internal(
+    AutoDisposeNotifierProvider<CurrentMapLoc, Coordinate>.internal(
   CurrentMapLoc.new,
   name: r'currentMapLocProvider',
   debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
@@ -21,6 +21,6 @@ final currentMapLocProvider =
   allTransitiveDependencies: null,
 );
 
-typedef _$CurrentMapLoc = AutoDisposeNotifier<LatLng>;
+typedef _$CurrentMapLoc = AutoDisposeNotifier<Coordinate>;
 // ignore_for_file: type=lint
-// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package
