@@ -22,4 +22,4 @@ final favoriteBeraProvider =
 
 typedef _$FavoriteBera = Notifier<List<AvalancheBulletin>>;
 // ignore_for_file: type=lint
-// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package
