@@ -6,12 +6,12 @@ part of 'favorite_station.dart';
 // RiverpodGenerator
 // **************************************************************************
 
-String _$favoriteStationHash() => r'588493fad97e9ae592f25ef30a42ab76aeba274f';
+String _$favoriteStationHash() => r'5e0800ede47b280c5084451f00b1c39d1454e4b5';
 
 /// See also [FavoriteStation].
 @ProviderFor(FavoriteStation)
 final favoriteStationProvider =
-    NotifierProvider<FavoriteStation, List<AbstractStation>>.internal(
+    AsyncNotifierProvider<FavoriteStation, List<AbstractStation>>.internal(
   FavoriteStation.new,
   name: r'favoriteStationProvider',
   debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
@@ -21,6 +21,6 @@ final favoriteStationProvider =
   allTransitiveDependencies: null,
 );
 
-typedef _$FavoriteStation = Notifier<List<AbstractStation>>;
+typedef _$FavoriteStation = AsyncNotifier<List<AbstractStation>>;
 // ignore_for_file: type=lint
-// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package
