@@ -6,7 +6,7 @@ part of 'station_data.dart';
 // RiverpodGenerator
 // **************************************************************************
 
-String _$stationDataHash() => r'ce443ccf193edba620865e676fdbbf4648a0618c';
+String _$stationDataHash() => r'fd3be9e3d19f28d57303c608d4944be9a00da4a9';
 
 /// See also [StationData].
 @ProviderFor(StationData)
@@ -22,4 +22,4 @@ final stationDataProvider =
 
 typedef _$StationData = AsyncNotifier<Map<int, List<DataStation>>>;
 // ignore_for_file: type=lint
-// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package
